requirements.txt
aniso8601
atomicwrites
attrs
certifi
charset-normalizer
click
colorama
Flask-cors
flask-login
flask-restful
Flask-Session
iniconfig
itsdangerous
Jinja2
MarkupSafe
more-itertools
packaging
psycopg2-binary
py
pyparsing
pytest
pytz
pyyaml
six
tomli
urllib3
wcwidth
Werkzeug